Add JetpackDashboardPage
