+++ projects/plugins/jetpack/tests/e2e/specs/connection/connection.test.js
@@ -1,6 +1,10 @@
 import { test, expect } from 'jetpack-e2e-commons/fixtures/base-test.js';
 import { doSiteLevelConnection, doClassicConnection } from 'jetpack-e2e-commons/flows/index.js';
-import { Sidebar, JetpackPage, DashboardPage } from 'jetpack-e2e-commons/pages/wp-admin/index.js';
+import {
+	Sidebar,
+	JetpackDashboardPage,
+	DashboardPage,
+} from 'jetpack-e2e-commons/pages/wp-admin/index.js';
 import { prerequisitesBuilder } from 'jetpack-e2e-commons/env/index.js';
 
 test.describe( 'Connection', () => {
@@ -28,7 +32,7 @@ test.describe( 'Connection', () => {
 		} );
 
 		await test.step( 'Can assert that site is connected', async () => {
-			const jetpackPage = await JetpackPage.visit( page );
+			const jetpackPage = await JetpackDashboardPage.visit( page );
 			expect( await jetpackPage.isConnected(), 'Site should be connected' ).toBeTruthy();
 		} );
 	} );
@@ -39,7 +43,7 @@ test.describe( 'Connection', () => {
 		} );
 
 		await test.step( 'Can assert that site is connected', async () => {
-			const jetpackPage = await JetpackPage.visit( page );
+			const jetpackPage = await JetpackDashboardPage.visit( page );
 			expect( await jetpackPage.isConnected(), 'Site should be connected' ).toBeTruthy();
 		} );
 	} );
