+++ tools/e2e-commons/pages/wp-admin/index.js
@@ -8,6 +8,7 @@ export { default as WordAdsBlock } from './blocks/word-ads.js';
 export { default as DashboardPage } from './dashboard.js';
 export { default as InPlacePlansPage } from './in-place-plans.js';
 export { default as JetpackPage } from './jetpack.js';
+export { default as JetpackDashboardPage } from './jetpack-dashboard.js';
 export { default as WPLoginPage } from './login.js';
 export { default as PluginsPage } from './plugins.js';
 export { default as ThemesPage } from './themes.js';
